\name{XLConnectJars-package}
\alias{XLConnectJars-package}
\alias{XLConnectJars}
\docType{package}
\title{
JAR dependencies for the XLConnect package
}
\description{
This package provides external JAR dependencies for the XLConnect package
}
\details{
\tabular{ll}{
Package: \tab XLConnectJars\cr
Type: \tab Package\cr
Version: \tab 0.1-0\cr
Date: \tab 2011-02-25\cr
URL: \tab \url{http://www.mirai-solutions.com}\cr
SystemRequirements: \tab java\cr
Depends: \tab R (>= 2.10.0), rJava\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
The dependency provided in this package is Apache POI 3.7.
}
\author{
Mirai Solutions GmbH, \email{xlconnect@mirai-solutions.com}
}
\references{
Mirai Solutions GmbH: \url{http://www.mirai-solutions.com}\cr
Apache POI: \url{http://poi.apache.org}
}
\keyword{ package }
