\name{XLConnectJars-package}
\alias{XLConnectJars-package}
\alias{XLConnectJars}
\docType{package}
\title{
JAR dependencies for the XLConnect package
}
\description{
This package provides external JAR dependencies for the XLConnect package
}
\details{
\tabular{ll}{
Package: \tab XLConnectJars\cr
Type: \tab Package\cr
Version: \tab 0.2-0\cr
Date: \tab 2012-07-18\cr
URL: \tab \url{http://www.mirai-solutions.com}\cr
\tab \url{http://miraisolutions.wordpress.com}\cr
SystemRequirements: \tab java (>= 6.0)\cr
Depends: \tab R (>= 2.10.0), rJava\cr
License: \tab GPL-3\cr
LazyLoad: \tab yes\cr
}
The dependency provided in this package is Apache POI 3.8.
}
\author{
Mirai Solutions GmbH, \email{xlconnect@mirai-solutions.com}
}
\references{
Mirai Solutions GmbH: \url{http://www.mirai-solutions.com}\cr
Mirai Solutions Blog: \url{http://miraisolutions.wordpress.com}\cr
Apache POI: \url{http://poi.apache.org}
}
\keyword{ package }
