\name{XLConnectJars-package}
\alias{XLConnectJars-package}
\alias{XLConnectJars}
\docType{package}
\title{
JAR dependencies for the XLConnect package
}
\description{
Provides external JAR dependencies for the XLConnect package
}
\details{
\tabular{ll}{
Package: \tab XLConnectJars\cr
Type: \tab Package\cr
Version: \tab 0.2-9\cr
Date: \tab 2014-08-13\cr
URL: \tab \url{http://www.mirai-solutions.com}\cr
\tab \url{http://miraisolutions.wordpress.com}\cr
SystemRequirements: \tab java (>= 1.6)\cr
Depends: \tab R (>= 2.10.0)\cr
Imports: \tab rJava\cr
License: \tab GPL-3\cr
Copyright: \tab See file COPYRIGHTS\cr
LazyLoad: \tab yes\cr
}
}
\author{
Mirai Solutions GmbH, \email{xlconnect@mirai-solutions.com}
}
\references{
Mirai Solutions GmbH: \url{http://www.mirai-solutions.com}\cr
Mirai Solutions on GitHub: \url{https://github.com/miraisolutions}\cr
Mirai Solutions Blog: \url{http://miraisolutions.wordpress.com}\cr
}
\keyword{ package }
