\name{XLConnectJars-package}
\alias{XLConnectJars-package}
\alias{XLConnectJars}
\docType{package}
\title{
JAR dependencies for the XLConnect package
}
\description{
Provides external JAR dependencies for the XLConnect package
}
\details{
\tabular{ll}{
Package: \tab XLConnectJars\cr
Type: \tab Package\cr
Version: \tab 0.2-14\cr
Date: \tab 2018-01-23\cr
URL: \tab \url{http://www.mirai-solutions.com}\cr
\tab \url{https://github.com/miraisolutions/xlconnect}\cr
BugReports: \tab \url{https://github.com/miraisolutions/xlconnect}\cr
SystemRequirements: \tab Java (>= 6)\cr
Imports: \tab rJava\cr
License: \tab GPL-3\cr
Copyright: \tab See file COPYRIGHTS\cr
LazyLoad: \tab yes\cr
}
}
\author{
Mirai Solutions GmbH, \email{xlconnect@mirai-solutions.com}
}
\references{
Mirai Solutions GmbH: \url{http://www.mirai-solutions.com}\cr
XLConnect on GitHub: \url{https://github.com/miraisolutions/xlconnect}\cr
Mirai Solutions on GitHub: \url{https://github.com/miraisolutions}
}
\keyword{ package }
